feat: Refactor agents for enhanced extensibility
