+++ multiAgentScenario.js
@@ -1,10 +1,9 @@
 const { StateGraph } = require("@langchain/langgraph");
-const { callGeminiCli } = require("./geminiCli");
+const { StateGraph } = require("@langchain/langgraph");
 const { z } = require("zod");
-const { Memory } = require("./memory");
 const config = require('./config');
-
-const memory = new Memory();
+const { agentOne } = require('./agents/agentOne');
+const { agentTwo } = require('./agents/agentTwo');
 
 // Define the state of our graph using Zod
 const graphState = {
@@ -22,50 +21,6 @@ const graphState = {
     type: "string",
   },
 };
-
-// Define the first agent
-async function preCallHook(agentName, state) {
-  console.log(`[Hook] Pre-call for ${agentName}. Current turn: ${state.turns}`);
-}
-
-async function postCallHook(agentName, state, response) {
-  console.log(`[Hook] Post-call for ${agentName}. Response length: ${response.length}`);
-}
-
-// Define the first agent
-async function agentOne(state) {
-  await preCallHook("Agent One", state);
-  const currentQuestion = state.question;
-  console.log("Agent One received question:", currentQuestion);
-  const response = await callGeminiCli(`As Agent One, provide an initial answer to: ${currentQuestion}`);
-  await postCallHook("Agent One", state, response);
-  await memory.saveMessage("Agent One", response);
-  return {
-    messages: state.messages.concat(`Agent One: ${response}`),
-    turns: state.turns + 1,
-    totalGeminiCalls: state.totalGeminiCalls + 1,
-    totalInputCharacters: state.totalInputCharacters + currentQuestion.length,
-    totalOutputCharacters: state.totalOutputCharacters + response.length,
-  };
-}
-
-// Define the second agent
-async function agentTwo(state) {
-  await preCallHook("Agent Two", state);
-  const lastMessage = state.messages[state.messages.length - 1];
-  const currentQuestion = state.question;
-  console.log("Agent Two received message:", lastMessage);
-  const response = await callGeminiCli(`As Agent Two, refine the following answer to "${currentQuestion}" based on this previous response: ${lastMessage}`);
-  await postCallHook("Agent Two", state, response);
-  await memory.saveMessage("Agent Two", response);
-  return {
-    messages: state.messages.concat(`Agent Two: ${response}`),
-    turns: state.turns + 1,
-    totalGeminiCalls: (state.totalGeminiCalls || 0) + 1,
-    totalInputCharacters: (state.totalInputCharacters || 0) + lastMessage.length + currentQuestion.length,
-    totalOutputCharacters: (state.totalOutputCharacters || 0) + response.length,
-  };
-}
 
 // Define a router to decide which agent to call next
 const MAX_TURNS = config.agents.maxTurns; // Define a maximum number of turns for collaboration
